refactor(examples): dedupe hold_key/wait/release_held_mouse_button in n2 adapters
CuaSandboxComputer (cua_adapter.py) and LocalX11Computer
(direct_x11_adapter.py) each hand-rolled three byte-identical async
methods relying only on each class's own key_down/key_up/left_mouse_up
and _left_mouse_down flag. Extracted PointerKeyLifecycleMixin onto
sandbox_tools.py, matching the existing ShellFileToolsMixin convention
(hook stubs the adapter already implements, mixed into the adapter
class). Internal helper, not re-exported from yutori/navigator/__init__.py
or documented in api.md, so no public API change.
